Refactor step 1
