+++ author.rb
@@ -1,0 +1,7 @@
+class Author
+  attr_reader :name, :biography
+  def initialize(name, biography)
+    @name = name
+    @biography = biography
+  end
+end
