+++ book.rb
@@ -1,0 +1,7 @@
+class Book
+  attr_reader :title, :author
+  def initialize(title, author)
+    @title = title
+    @author = author
+  end
+end
