+++ order.rb
@@ -1,0 +1,8 @@
+class Order
+  attr_reader :book, :reader, :date
+  def initialize(book, reader, date)
+    @book = book
+    @reader = reader
+    @date = date
+  end
+end
